Dataclasses and started splitting up code
