+++ workstate/exceptions.py
@@ -1,0 +1,5 @@
+'''WorkState exceptions'''
+
+
+class BrokenStateModelException(Exception):
+    '''State model is broken'''
